fixed null items on stock error

+++ app/src/main/java/com/ap/mobile/stocks/data/local/entity/chart/TodayChart.kt
@@ -53,13 +53,13 @@ data class TodayChart(
 	val low: Double? = null,
 
 	@field:SerializedName("changeOverTime")
-	val changeOverTime: Int? = null,
+	val changeOverTime: Double? = null,
 
 	@field:SerializedName("numberOfTrades")
 	val numberOfTrades: Int? = null,
 
 	@field:SerializedName("marketChangeOverTime")
-	val marketChangeOverTime: Int? = null,
+	val marketChangeOverTime: Double? = null,
 
 	@field:SerializedName("close")
 	val close: Double? = null,
